+++ hutool-http/src/test/java/cn/hutool/http/HttpUtilTest.java
@@ -378,4 +378,9 @@ public class HttpUtilTest {
 		final HttpRequest request = HttpRequest.of(url).method(Method.GET);
 		Console.log(request.execute().body());
 	}
+
+	@Test
+	public void allowPatchTest() {
+		HttpGlobalConfig.allowPatch();
+	}
 }